CourtSearch/requirements.txt
# Court Lookup Script Dependencies
# Python 3.10+ required

requests>=2.31.0          # HTTP requests for web scraping
beautifulsoup4>=4.12.0    # HTML parsing
lxml>=4.9.0               # Fast HTML/XML parser backend
playwright>=1.40.0        # JavaScript rendering backup
tabulate>=0.9.0           # Clean table output formatting
python-dateutil>=2.8.0    # Date parsing utilities
playwright-stealth>=1.0.6 # Stealth mode for Playwright (anti-bot evasion)
